+++ app/assets/javascripts/govuk_publishing_components/lib/trigger-event.js
@@ -4,15 +4,26 @@
 
   window.GOVUK.triggerEvent = function (element, eventName, parameters) {
     var params = parameters || {}
-    params.bubbles = true
-    params.cancelable = true
     var event
+    var keyCode = params.keyCode
+
+    if (!Object.prototype.hasOwnProperty.call(params, 'bubbles')) {
+      params.bubbles = true
+    }
+
+    if (!Object.prototype.hasOwnProperty.call(params, 'cancelable')) {
+      params.cancelable = true
+    }
 
     if (typeof window.CustomEvent === 'function') {
       event = new window.CustomEvent(eventName, params)
     } else {
       event = document.createEvent('CustomEvent')
       event.initCustomEvent(eventName, params.bubbles, params.cancelable, params.detail)
+    }
+
+    if (keyCode) {
+      event.keyCode = keyCode
     }
 
     element.dispatchEvent(event)